+++ packages/core/src/invocation.js
@@ -81,6 +81,22 @@ class IssuedInvocation {
     this.notBefore = notBefore
     this.nonce = nonce
     this.facts = facts
+
+    /** @type {API.BlockStore<unknown>} */
+    this.attachedBlocks = new Map()
+  }
+
+  /**
+   * Attach a block to the invocation DAG so it would be included in the
+   * block iterator.
+   * ⚠️ You should only attach blocks that are referenced from the `capabilities`
+   * or `facts`, if that is not the case you probably should reconsider.
+   * ⚠️ Once a delegation is de-serialized the attached blocks will not be re-attached.
+   *
+   * @param {API.Block} block
+   */
+  attach(block) {
+    this.attachedBlocks.set(`${block.cid}`, block)
   }
 
   delegate() {